# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Analyzer and Job Matcher API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type ResumeStatus = (typeof ResumeStatus)[keyof typeof ResumeStatus];

export const ResumeStatus = {
  pending: "pending",
  analyzing: "analyzing",
  completed: "completed",
  failed: "failed",
} as const;

export interface Resume {
  id: number;
  userId: string;
  filename: string;
  originalName: string;
  status: ResumeStatus;
  /** @nullable */
  atsScore?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ExperienceItem {
  title: string;
  company: string;
  duration: string;
  description: string;
}

export interface ResumeAnalysis {
  id: number;
  resumeId: number;
  skills: string[];
  experience: ExperienceItem[];
  keywords: string[];
  atsScore: number;
  improvements: string[];
  summary: string;
  createdAt: string;
}

export interface ResumeWithAnalysis {
  resume: Resume;
  /** @nullable */
  analysis?: ResumeAnalysis | null;
}

export interface Job {
  id: number;
  title: string;
  company: string;
  location: string;
  category: string;
  description: string;
  requiredSkills: string[];
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  postedAt: string;
}

export interface JobMatch {
  job: Job;
  matchPercentage: number;
  matchedSkills: string[];
  missingSkills: string[];
}

export interface DashboardStats {
  totalResumes: number;
  /** @nullable */
  averageAtsScore?: number | null;
  topSkills: string[];
  recentResumes: Resume[];
  /** @nullable */
  bestJobMatchScore?: number | null;
}

export type UploadResumeBody = {
  file: Blob;
};

export type GetJobsParams = {
  search?: string;
  category?: string;
  /**
   * @nullable
   */
  resumeId?: number | null;
};