# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Resume Analyzer and Job Matcher API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: resumes
    description: Resume operations
  - name: jobs
    description: Job operations
  - name: dashboard
    description: Dashboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /resumes:
    get:
      operationId: getResumes
      tags: [resumes]
      summary: List all resumes for the authenticated user
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /resumes/upload:
    post:
      operationId: uploadResume
      tags: [resumes]
      summary: Upload a resume PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Resume uploaded and queued for analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /resumes/{resumeId}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a specific resume with full analysis
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume with analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeWithAnalysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /resumes/{resumeId}/analyze:
    post:
      operationId: analyzeResume
      tags: [resumes]
      summary: Trigger AI analysis of a resume
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: Search and filter jobs
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: resumeId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a specific job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/matches/{resumeId}:
    get:
      operationId: getJobMatches
      tags: [jobs]
      summary: Get jobs matched to a resume with percentage scores
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matched jobs with scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobMatch"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics for the authenticated user
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    Resume:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        filename:
          type: string
        originalName:
          type: string
        status:
          type: string
          enum: [pending, analyzing, completed, failed]
        atsScore:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - filename
        - originalName
        - status
        - createdAt
        - updatedAt

    ResumeAnalysis:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        skills:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            $ref: "#/components/schemas/ExperienceItem"
        keywords:
          type: array
          items:
            type: string
        atsScore:
          type: integer
        improvements:
          type: array
          items:
            type: string
        summary:
          type: string
        createdAt:
          type: string
      required:
        - id
        - resumeId
        - skills
        - experience
        - keywords
        - atsScore
        - improvements
        - summary
        - createdAt

    ExperienceItem:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        duration:
          type: string
        description:
          type: string
      required:
        - title
        - company
        - duration
        - description

    ResumeWithAnalysis:
      type: object
      properties:
        resume:
          $ref: "#/components/schemas/Resume"
        analysis:
          type: ["object", "null"]
          $ref: "#/components/schemas/ResumeAnalysis"
      required:
        - resume

    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        location:
          type: string
        category:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        postedAt:
          type: string
      required:
        - id
        - title
        - company
        - location
        - category
        - description
        - requiredSkills
        - postedAt

    JobMatch:
      type: object
      properties:
        job:
          $ref: "#/components/schemas/Job"
        matchPercentage:
          type: integer
        matchedSkills:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
      required:
        - job
        - matchPercentage
        - matchedSkills
        - missingSkills

    DashboardStats:
      type: object
      properties:
        totalResumes:
          type: integer
        averageAtsScore:
          type: ["integer", "null"]
        topSkills:
          type: array
          items:
            type: string
        recentResumes:
          type: array
          items:
            $ref: "#/components/schemas/Resume"
        bestJobMatchScore:
          type: ["integer", "null"]
      required:
        - totalResumes
        - topSkills
        - recentResumes